PDI download update

+++ fw/AMC/src/osal/src/freeRTOS/osal.c
@@ -1954,7 +1954,7 @@ void vOSAL_ExitCritical( void )
 /**
  * @brief   OSAL wrapper for task/thread safe memory allocation.
  */
-void* pvOSAL_MemAlloc( uint16_t xSize )
+void* pvOSAL_MemAlloc( uint32_t xSize )
 {
     void* pvMemory = NULL;
 
@@ -1963,12 +1963,12 @@ void* pvOSAL_MemAlloc( uint16_t xSize )
         if( TRUE == iOsStarted )
         {
             /* thread safe */
-            pvMemory = pvPortMalloc( ( size_t )xSize );
+            pvMemory = pvPortMalloc( xSize );
         }
         else
         {
             /* note: Not thread safe */
-            pvMemory = malloc( ( size_t )xSize );
+            pvMemory = malloc( xSize );
         }
 
         pxOsStatsHandle->iMemAllocCallCount++;
@@ -1979,7 +1979,7 @@ void* pvOSAL_MemAlloc( uint16_t xSize )
 /**
  * @brief   OSAL wrapper for task/thread safe memory set.
  */
-void* pvOSAL_MemSet( void* pvDestination, int iValue, uint16_t usSize )
+void* pvOSAL_MemSet( void* pvDestination, int iValue, uint32_t usSize )
 {
     void* pvSetMemory = NULL;
 
@@ -1993,7 +1993,7 @@ void* pvOSAL_MemSet( void* pvDestination, int iValue, uint16_t usSize )
             if( pdPASS == xSemaphoreTakeFromISR( ( SemaphoreHandle_t ) pvMemSetSemHandle, &xHigherPriorityTaskWoken ) )
             {
                 /* semaphore taken successfully, set memory */
-                pvSetMemory = memset( pvDestination, iValue, ( size_t )usSize );
+                pvSetMemory = memset( pvDestination, iValue, usSize );
 
                 /* release semaphore */
                 xSemaphoreGiveFromISR( ( SemaphoreHandle_t ) pvMemSetSemHandle, &xHigherPriorityTaskWoken );
@@ -2002,7 +2002,7 @@ void* pvOSAL_MemSet( void* pvDestination, int iValue, uint16_t usSize )
         else
         {
             /* set memory, note: Not thread safe */
-            pvSetMemory = memset( pvDestination, iValue, ( size_t )usSize );
+            pvSetMemory = memset( pvDestination, iValue, usSize );
         }
     }
     return pvSetMemory;
@@ -2011,7 +2011,7 @@ void* pvOSAL_MemSet( void* pvDestination, int iValue, uint16_t usSize )
 /**
  * @brief   OSAL wrapper for task/thread safe memory copy.
  */
-void* pvOSAL_MemCpy( void* pvDestination, const void* pvSource, uint16_t usSize )
+void* pvOSAL_MemCpy( void* pvDestination, const void* pvSource, uint32_t usSize )
 {
     void* pvSetMemory = NULL;
 
@@ -2026,7 +2026,7 @@ void* pvOSAL_MemCpy( void* pvDestination, const void* pvSource, uint16_t usSize 
             if( pdPASS == xSemaphoreTakeFromISR( ( SemaphoreHandle_t ) pvMemCpySemHandle, &xHigherPriorityTaskWoken ) )
             {
                 /* semaphore taken successfully, copy memory */
-                pvSetMemory = memcpy( pvDestination, pvSource, ( size_t )usSize );
+                pvSetMemory = memcpy( pvDestination, pvSource, usSize );
 
                 /* release semaphore */
                 xSemaphoreGiveFromISR( ( SemaphoreHandle_t ) pvMemCpySemHandle, &xHigherPriorityTaskWoken );
@@ -2035,7 +2035,7 @@ void* pvOSAL_MemCpy( void* pvDestination, const void* pvSource, uint16_t usSize 
         else
         {
             /* copy memory, note: Not thread safe */
-            pvSetMemory = memcpy( pvDestination, pvSource, ( size_t )usSize );
+            pvSetMemory = memcpy( pvDestination, pvSource, usSize );
         }
     }
     return pvSetMemory;
@@ -2044,7 +2044,7 @@ void* pvOSAL_MemCpy( void* pvDestination, const void* pvSource, uint16_t usSize 
 /**
  * @brief   OSAL wrapper for task/thread safe memory movement.
  */
-void vOSAL_MemMove( void *pvDestination, void *pvSource, uint16_t usPayload_size )
+void vOSAL_MemMove( void *pvDestination, void *pvSource, uint32_t usPayload_size )
 {
     if( ( NULL != pvDestination ) &&
         ( NULL != pvSource ) )
@@ -2059,7 +2059,7 @@ void vOSAL_MemMove( void *pvDestination, void *pvSource, uint16_t usPayload_size
                 /* semaphore taken successfully, move memory */
                 memmove( pvDestination,
                          pvSource,
-                         ( size_t )usPayload_size );
+                         usPayload_size );
                 /* release semaphore */
                 xSemaphoreGiveFromISR( ( SemaphoreHandle_t ) pvMemMoveSemHandle, &xHigherPriorityTaskWoken );
             }
@@ -2069,7 +2069,7 @@ void vOSAL_MemMove( void *pvDestination, void *pvSource, uint16_t usPayload_size
             /* note: Not thread safe */
             memmove( pvDestination,
                      pvSource,
-                     ( size_t )usPayload_size );
+                     usPayload_size );
         }
     }
 }
@@ -2172,7 +2172,7 @@ char cOSAL_GetChar( void )
 /**
  * @brief   OSAL wrapper for task/thread safe string copy.
  */
-char* pcOSAL_StrNCpy( char *pcDestination, const char *pcSource, uint16_t usSize )
+char* pcOSAL_StrNCpy( char *pcDestination, const char *pcSource, uint32_t usSize )
 {
     char *pcSetString = NULL;
 
@@ -2188,7 +2188,7 @@ char* pcOSAL_StrNCpy( char *pcDestination, const char *pcSource, uint16_t usSize
             if ( pdPASS == xSemaphoreTakeFromISR( ( SemaphoreHandle_t ) pvStrNCpySemHandle, &xHigherPriorityTaskWoken ) )
             {
                 /* semaphore taken successfully, copy string */
-                pcSetString = strncpy( pcDestination, pcSource, ( size_t )usSize );
+                pcSetString = strncpy( pcDestination, pcSource, usSize );
 
                 /* release semaphore */
                 xSemaphoreGiveFromISR( ( SemaphoreHandle_t ) pvStrNCpySemHandle, &xHigherPriorityTaskWoken );
@@ -2197,7 +2197,7 @@ char* pcOSAL_StrNCpy( char *pcDestination, const char *pcSource, uint16_t usSize
         else
         {
             /* copy string, note: Not thread safe */
-            pcSetString = strncpy( pcDestination, pcSource, ( size_t )usSize );
+            pcSetString = strncpy( pcDestination, pcSource, usSize );
         }
     }
     return pcSetString;
@@ -2206,7 +2206,7 @@ char* pcOSAL_StrNCpy( char *pcDestination, const char *pcSource, uint16_t usSize
 /**
  * @brief   OSAL wrapper for task/thread safe memory compare.
  */
-int iOSAL_MemCmp( const void *pvMemoryOne, const void *pvMemoryTwo, uint16_t usSize )
+int iOSAL_MemCmp( const void *pvMemoryOne, const void *pvMemoryTwo, uint32_t usSize )
 {
     int iStatus = ERROR;
 
@@ -2222,7 +2222,7 @@ int iOSAL_MemCmp( const void *pvMemoryOne, const void *pvMemoryTwo, uint16_t usS
             if ( pdPASS == xSemaphoreTakeFromISR( ( SemaphoreHandle_t ) pvMemCmpSemHandle, &xHigherPriorityTaskWoken ) )
             {
                 /* semaphore taken successfully, compare memory */
-                iStatus = memcmp( pvMemoryOne, pvMemoryTwo, ( size_t )usSize );
+                iStatus = memcmp( pvMemoryOne, pvMemoryTwo, usSize );
 
                 /* release semaphore */
                 xSemaphoreGiveFromISR( ( SemaphoreHandle_t ) pvMemCmpSemHandle, &xHigherPriorityTaskWoken );
@@ -2231,7 +2231,7 @@ int iOSAL_MemCmp( const void *pvMemoryOne, const void *pvMemoryTwo, uint16_t usS
         else
         {
             /* compare memory, note: Not thread safe */
-            iStatus = memcmp( pvMemoryOne, pvMemoryTwo, ( size_t )usSize );
+            iStatus = memcmp( pvMemoryOne, pvMemoryTwo, usSize );
         }
     }
     return iStatus;
